pyproject.toml
[project]
name = "kai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.112.2",
    "exceptiongroup>=1.2.2",
    "typesense>=0.19.0",
    "ruff>=0.6.2",
    "pydantic-settings>=2.4.0",
    "SQLAlchemy>=2.0.32",
    "uvicorn>=0.30.6",
    "black>=24.8.0",
    "sql-metadata>=2.13.0",
    "cryptography>=43.0.0",
    "langchain>=1.0.2,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-community>=0.4.0,<1.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    "langchain-ollama>=1.0.0,<2.0.0",
    "langchain-google-genai>=3.0.0,<4.0.0",
    "langgraph>=1.0.0,<2.0.0",
    "langgraph-prebuilt>=0.1.1",
    "deepagents>=0.2.8",
    "overrides>=7.7.0",
    "pandas>=2.2.2",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.9",
    "llama-index>=0.13.0",
    "llama-index-vector-stores-typesense>=0.4.0",
    "pytest>=8.3.3",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.24.0",
    "pyarrow>=19.0.0",
    "llama-index-llms-langchain>=0.7.0",
    "llama-index-embeddings-langchain>=0.4.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "matplotlib>=3.8.0",
    "openpyxl>=3.1.0",
    "temporalio>=1.4.0",
    "httpx>=0.27.0",
    "neo4j>=6.0.3",
    "plotly>=5.0.0",
    "jsonschema>=4.0.0",
    "scipy>=1.11.0",
    # Security patches for transitive dependencies
    "urllib3>=2.6.0", # CVE: streaming API and decompression chain vulnerabilities
    "marshmallow>=3.26.2", # CVE: DoS in Schema.load(many)
    "psycopg2-binary>=2.9.11",
    "google-genai>=1.62.0",
]

[project.scripts]
kai = "app.modules.autonomous_agent.main:cli"
kai-analytics = "app.modules.analytics.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
